Module Gamelle.Box

Axis-aligned bounding boxes.

type t = Gg.box2

The type of axis-aligned bounding boxes (rectangles without a rotation).

val v : Point.t -> Size.t -> t

v topleft size is a box with origin point topleft and dimension size.

val zero : t

zero is the empty box located at the origin.

Draw

val draw : io:io -> ?color:Color.t -> t -> unit

draw ~io b draws the outline border of the box b.

val fill : io:io -> ?color:Color.t -> t -> unit

fill ~io b fills the inside of the box b.

Accessors

val o : t -> Point.t

o b returns the top-left point of the box b.

val size : t -> Size.t

size b returns the dimension of the box b.

val x_left : t -> float

x_left b is the left-most x coordinate of the box b.

val x_middle : t -> float

x_middle b is the center x coordinate of the box b.

val x_right : t -> float

x_right b is the right-most x coordinate of the box b.

val y_top : t -> float

y_top b is the upper y coordinate of the box b.

val y_middle : t -> float

y_middle b is the center y coordinate of the box b.

val y_bottom : t -> float

y_bottom b is the lower y coordinate of the box b.

val left : t -> Segment.t

left b returns the left segment of the box b.

val right : t -> Segment.t

right b returns the right segment of the box b.

val top : t -> Segment.t

top b returns the top segment of the box b.

val bottom : t -> Segment.t

bottom b returns the bottom segment of the box b.

val mem : Point.t -> t -> bool

mem pt b returns true if the point pt lies inside the box b, false otherwise.

val random_mem : t -> Point.t

random_mem b returns a random point inside the box b.