Module Gamelle.Circle

Circles.

type t

The type of circles.

val v : Point.t -> float -> t

v pt r is a circle with center point pt and radius r.

Draw

val draw : io:io -> ?color:Color.t -> t -> unit

draw ~io c draws the outline border of the circle c.

val fill : io:io -> ?color:Color.t -> t -> unit

fill ~io c fills the inside of the circle c.

Accessors

val center : t -> Point.t

center c is the center point of the circle c.

val radius : t -> float

radius c is the radius of the circle c.

Transforms

val translate : Vec.t -> t -> t

translate v c translates the circle c by vector v.

scale f c scales the circle c by factor f.

Collisions

val mem : Point.t -> t -> bool

mem pt c returns true if point pt lies inside the circle c, false otherwise.

val intersect : t -> t -> bool

intersect a b returns true if circle a intersects circle b, false otherwise.

val intersections : t -> t -> Point.t list

intersections a b returns the list of intersections points between the circles a and b.