Module Gamelle.Input

Player inputs: mouse and keyboard events.

type key = [

| `alt

| `alt_gr

| `arrow_down

| `arrow_left

| `arrow_right

| `arrow_up

| `backspace

| `click_left

| `click_right

| `control_left

| `control_right

| `delete

| `escape

| `input_char of string

| `meta

| `physical_char of char

| `quit

| `shift

| `space

| `tab

| `wheel

| `unknown_key

]

The type of player inputs.

val is_pressed : io:io -> key -> bool

is_pressed ~io key returns true if the player is currently holding key.

val is_down : io:io -> key -> bool

is_down ~io key returns true if the player just started pressing the key.

val is_up : io:io -> key -> bool

is_up ~io key returns true if the player was holding key and just released it.

Mouse

val mouse_pos : io:io -> Point.t

mouse_pos ~io returns the coordinates of the player mouse.

val wheel_delta : io:io -> float

wheel_delta ~io returns the amount of change of the mouse wheel.