Module Gamelle.Physics

Rigid physics for Shape objects.

type t

The type of rigid bodies.

type kind =

| Movable

| Immovable

val v : ?mass:float -> ?inertia:float -> ?restitution:float -> ?kind:kind -> Shape.t -> t

v s returns a new rigid body with shape s.

?mass is the weight of the rigid body.

?inertia controls the reactivity of the rigid body.

?restitution defines the bouncyness of the rigid body.

?kind controls if the rigid body can move.

Accessors

val center : t -> Point.t

Simulation

val add_velocity : Vec.t -> t -> t

add_velocity v t adds v to the rigid body t current velocity.

val add_rot_velocity : float -> t -> t

add_rot_velocity r t adds r to the rigid body t current rotational velocity.

val update : dt:float -> t -> t

update ~dt t updates the rigid body t position and rotation according to its current velocities and delta time dt.

val fix_collisions : t list -> t list

fix_collisions lst detects and repairs any collisions between the rigid bodies in the list lst.

Draw

val draw : io:io -> ?color:Color.t -> t -> unit

draw ~io t draws the outline border of the rigid body t.

val fill : io:io -> ?color:Color.t -> t -> unit

fill ~io t fills the inside of the rigid body t.