Module Gamelle.Point

Points: x and y positions in 2D.

type t = Gg.p2

The type of positions.

val v : float -> float -> t

v x y returns a point at coordinates x and y.

val o : t

o is the point at the zero origin v 0.0 0.0

Accessors

val x : t -> float

x pt returns the x coordinate of the point pt.

val y : t -> float

y pt returns the y coordinate of the point pt.

val lerp : float -> t -> t -> t

lerp t a b is the linear interpolation between the two points a and b controlled by the time parameter t.

lerp 0.0 a b = a

lerp 1.0 a b = b

lerp 0.5 a b = mid point of a and b