Module Gamelle.Polygon

Polygons.

type t

The type of polygons.

val v : Point.t list -> t

v pts is a polygon with vertices pts.

val draw : io:io -> ?color:Color.t -> t -> unit

draw ~io p draws the outline border of the polygon p.

val fill : io:io -> ?color:Color.t -> t -> unit

fill ~io p fills the inside of the polygon p.

Accessors

val center : t -> Point.t

center p returns the center of mass of the polygon p.

val points : t -> Point.t list

segments p returns the list of vertices of the polygon p.

val segments : t -> Segment.t list

segments p returns the list of segments of the polygon p.

val bounding_box : t -> Box.t

bounding_box p returns the bounding box of the polygon p.

Transforms

val translate : Vec.t -> t -> t

translate v p translates the polygon p by vector v.