Module Gamelle.Text

Text rendering.

val draw : io:io -> ?color:Color.t -> ?font:Font.t -> ?size:int -> at:Point.t -> string -> unit

draw ~io ~at txt prints the string txt at position at on the screen.

val draw_multiline : io:io -> ?color:Color.t -> ?width:float -> ?interline:float -> ?font:Font.t -> ?size:int -> at:Point.t -> string -> unit

draw_multiline ~io ~at txt prints the string txt at position at on the screen, possibly wrapping it on multiple lines if the text overflows ?width.

Measure

val size : io:io -> ?font:Font.t -> ?size:int -> string -> Size.t

size ~io str returns the size that would be used by draw to render the text str.

val size_multiline : io:io -> ?width:float -> ?interline:float -> ?font:Font.t -> ?size:int -> string -> Size.t

size_multiline ~io str returns the size that would be used by draw_multiline to render the text str on multiple lines.