Module Gamelle.Vec

Vectors: directions in 2D.

type t = Gg.v2

The type of directional vectors.

val v : float -> float -> t

v x y is a vector with direction x, y.

val zero : t

zero is the zero vector v 0.0 0.0.

Accessors

val x : t -> float

x v is the x component of the vector v.

val y : t -> float

y v is the y component of the vector v.

val to_tuple : t -> float * float

to_tuple v is x v, y v.

Linear algebra

val norm : t -> float

norm v is the length of the vector v.

val unit : t -> t

unit v is the vector v normalized to have a unit length.

val dot : t -> t -> float

dot a b is the dot product of the vectors a, b.

Operators

val (*) : float -> t -> t

f * v is the vector v scaled by a factor f.

val (+) : t -> t -> t

a + b is the sum of the two vectors a, b.

val (-) : t -> t -> t

a - b is the substraction of vector a by b.