Module Gamelle.View

Customize the io camera, default color and font used.

Viewport

val translate : Vec.t -> io -> io

translate v io translates everything by vector v.

val scale : float -> io -> io

scale f io scales everything by a factor f.

val rotate : float -> io -> io

rotate a io rotates everything by angle a.

val drawing_box : ?scale:bool -> ?set_window_size:bool -> Box.t -> io -> io

drawing_box b io ensures the box b matches the dimensions of the window. See Window.size.

val clip : Box.t -> io -> io

clip b io ensures no drawing can happen outside of the box b.

val z_index : int -> io -> io

z_index z io controls the depth z of the following draws.

val color : Color.t -> io -> io

color c io specifies the default color c.

Font

val font : Font.t -> io -> io

font f io specifies the default font f.

val font_size : int -> io -> io

font_size s io specifies the default font size s.