Module Gamelle.Window

Configure the game window.

val show_cursor : io:io -> bool -> unit

show_cursor ~io visible toggles the visibility of the operating system mouse cursor.

val set_size : io:io -> Size.t -> unit

set_size ~io wh resizes the operating system window.

Example:

Window.set_size (Size.v 800. 800.) ;

val size : io:io -> Size.t

size ~io returns the current size of the game window.

val box : io:io -> Box.t

box ~io returns a box matching the game window contents.

Example:

(* clear the screen *)
Box.fill ~io ~color:Color.white (Window.box ~io) ;